pyproject.toml
[tool.poetry]
name = "telegram_bot_scripts"
version = "0.1.0"
description = ""
authors = ["Ian Scott Knight <isk@alumni.stanford.edu>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.25.1"
python-dotenv = "^0.15.0"
schedule = "^1.0.0"
pyicloud = "^0.10.2"
icalevents = "^0.1.25"

[tool.poetry.dev-dependencies]
pre-commit = "^2.11.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
